This is a fix for #83 support @Rule on methods Extend rules to support @Rule public MethodRule someRule() { return new SomeRule(); }
Scala does not allow the creation of public fields. If you declare a field public, the field is created as private but with accessor methods.
To allow the use of @Rule with Scala fields, this patch allows the addition
of @Rule to methods. When BlockJUnit4ClassRunner retrieves the list of fields
that have the @Rule annotation, it also searches for methods with @Rule.
These methods are called and the values returned are added to the list of
TestRule.

@ClassRule has been similarly changed.

This is intended to be used as follows:

private ExpectedException thrown = rules.ExpectedException.none();

@Rule public ExpectedException getThrown() {
    return thrown;
}

or

@Rule public Timeout getTimeout() {
    return new Timeout(20);
}

The validation is exactly the same as for the fields, i.e the methods must be
public and return a TestRule/MethodRule. In the case of @ClassRule, the
methods must also be static.