This is the implementation for a feature of Configurable Categories https://github.com/KentBeck/junit/issues/307
You can make a narrow list of tests in your build process according to build specifics.
See the test CategoriesConfiguredBySystemPropertiesTest.
There the suite accepts almost all tests of platforms, and the system properties suppress some of them.

Here we use a Set of categories which was necessary in this feature. Not using single values in variables fIncluded either fExcluded in Categories class. See their comments in code.

This way we have built also a prerequisite for #336. See 'TODO' and 'FIX for issue #336' in class Categories.