Adding support for marking a Test as blocked.
Previously there were only two expected outcomes
from a Test, passed or failed. If some precondition was not met
that was not part of the actual test it could only by failed.
With this patch user can now call:

Assert.block() or Assert.block(message) to stop execution of test
and mark it as blocked. The individual test result will now be a B
instead of F. The overall summary will also show the total number
of blocked Tests.