Introduce withStacktrace() matcher
StacktracePrintingMatcher delegates to a Matcher<Throwable> and in
addition appends the stacktrace of the actual item in case of a
mismatch.

It is used inside ExpectedExceptionMatcher and can also be used
separately when matching Throwables by calling `assertThat(actual,
withStacktrace(throwableMatcher)`.

assertThat no longer sets the cause of the AssertionError. CauseMatcher
has been removed.