Implement test cases for the RANDOM sorting implementations.
These tests need improvement so that they pass with higher probability.
It is a good start however and is good for seeing the level of bias
that exists with the random comparator. It will also show how terribly
slow the random comparator is as compared to using
Collections.shuffle().

As far as improving the tests, there is a book mentioned in Volume 2 of
The Art of Computer Programming, titled "How to Lie With Statistics" by
Darrell Huff (Norton, 1954). Information in this book may be useful in
improving these tests.