Disable doclint for all groups but accessibility and reference.
With JDK 8, doclint has been added to Javac and Javadoc. It is enabled
by default but it is too strict for JUnit. The only groups that don't
report issues are accessibility, reference and syntax. But syntax has
to be disabled for Javadoc because the Hamcrest classes are reporting
syntax issues.

Additionally doclint is disabled for JDK < 8, because the -Xdoclint
argument causes a failing build in JDKs before JDK 8.