# This is a combination of 10 commits. # The first commit's message is: Remove outdated documentation.
# This is the 2nd commit message:

Fix Javadoc issues reported by JDK8's doclint.

# This is the 3rd commit message:

Disable doclint for all groups but accessibility and reference.

With JDK 8, doclint has been added to Javac and Javadoc. It is enabled
by default but it is too strict for JUnit. The only groups that don't
report issues are accessibility, reference and syntax. But syntax has
to be disabled for Javadoc because the Hamcrest classes are reporting
syntax issues.

Additionally doclint is disabled for JDK < 8, because the -Xdoclint
argument causes a failing build in JDKs before JDK 8.

# This is the 4th commit message:

Verify javadoc on Travis CI.

The build should fail for Javadoc issues.

# This is the 5th commit message:

Use compilerArgs instead of compilerArguments.

The parameter compilerArguments is deprecated.

# This is the 6th commit message:

Add additional Javadoc for TestFailure.java
# This is the 7th commit message:

Add additional Javadoc for AssertionFailedError.java
# This is the 8th commit message:

Explicate javadoc further
# This is the 9th commit message:

Fix mistakes and format in Javadoc

# This is the 10th commit message:

Fix formatting issues in javadoc