Remove redundant logic from DescriptionBuilder
Current implementation is based on the approach usually used to enforce at least one argument on varargs method invocation. In given context such an enforcement has no meaning as invoking withAdditionalAnnotations() with no arguments is the same nonsense for the author as invoking withAdditionalAnnotations(null). There are also no any null-checking and the consequences of both forms are the same.  Motivation is to remove redundant instance construction and some CPU cycles.