Do not include thread start time in test timeout measurements.
The time it takes to start a thread can be surprisingly large.
Especially in virtualized cloud environments where noisy neighbours.
This change reduces the probability of non-deterministic failures of
tests with timeouts (`@Test(timeout=…)`) by not beginning the timeout
clock until we have observed the starting of the task thread – the
thread that runs the actual test. This should make tests with small
timeout values more reliable in general, and especially in cloud CI
environments.

No tests added for this because of the probabilistic nature of the
problem, and the small scope of the change.