FailOnTimeoutTest: Make assertion in thread test
The test added in 8d7a917006cac1669b914c6ce6ec98a2d3d3c074 simply
printed out the stack trace--it was not an automated test, but instead
required one to manually check the output.

Replace that test with one which makes an assertion: a thread with the
specified description is created to evaluate the FailOnTimeout.