Timeout: Deprecate createFailOnTimeoutStatement
In order to give timeout test threads descriptive names, we want to
change the createFailOnTimeoutStatement API to take a `Description`. We
can't change the old method, however, as it's a protected API.

Deprecate the old method, and have a new private method with a
`Description` parameter call the old. The old method will be removed
in a future version.