Suggested improvement - use the TestClass object to create instances of the test object. This allows for a runner to be created by something which intends to modify the test object before the test starts. This can lead to one runner executing a different sort of runner within an inner class of the TestClass, passing context to the inner class's object.
Example - a MockitoJunitRunner could be used to inject values into
a test object running in a SpringJUnitRunner inner class. This would
require those runners to change too, of course.

Right now, the problem with runners is that they're monolithic and use
inheritance to share features. This change is intended to allow more
injection of behaviour/data across runners via the test object.