Updated answer to question in faq.html
Updated answer as discussed on JUnit's Yahoo! Group:

If you are using JDK 1.3 or higher, you can use reflection to subvert the access control mechanism with the aid of the <a href="http://sourceforge.net/projects/privaccessor/">PrivilegedAccessor</a>. For details on how to use it, read <a href="http://www.onjava.com/pub/a/onjava/2003/11/12/reflection.html">this article</a>.
If you are using JDK 1.6 or higher and you annotate your tests with @Test, you can use <a href="http://dp4j.com">Dp4j</a> to inject reflection in your
test methods. For details on how to use it, see <a href="http://dp4j.com/testscript">this test script</a>.