Add tests that were not previously being run
The TestWatchmanTest failed because it was using @Rule on a
static field. Switching to @ClassRule did not work because it
does not implement TestRule. I fixed the test by making the
fields non-static however presumably this test did work at some
point which means that there could have been a regression.

After this change 935 tests passed, 7 were ignored.