Add assertThrows overloads with ThrowingSupplier
This commit aligns `Assert` with Jupiter's `Assertions` class by adding
overloaded variants of `assertThrows` that take a `ThrowingSupplier`
instead of a `ThrowingRunnable` which was originally proposed and
implemented by @cushon.

If the given `ThrowingSupplier` fails to throw the expected exception
and instead returns a value, the string representation of the value is
included in the failure message to aide debugging.