Fix "java"-based tests adding well-known class-path elements
Prior to this commit the value of `System.getProperty("java.class.path")`
and the resolution of `getClass().getClassLoader().getResource(".")` was
at test execution was able to provide the needed class-path elements for
forking an external JVM.
Using the DIRECT execution mode of "junit-platform-maven-plugin" does no
longer leak that information to tests.

Instead of hard-coding "target/classes" and "target/test-classes" like
done is commit, the values may be configured and passed as special
runtime properties.