Deprecate Assert#assertThat
The method "assertThat" is used for writing assertions with Hamcrest.
Hamcrest is an independent assertion library and contains an own
"assertThat" method in the class "org.hamcrest.MatcherAssert". It is
available both in the old Hamcrest 1.3 release and in the current
Hamcrest 2.1. Therefore the JUnit team recommends to use Hamcrest's own
"assertThat" directly.