Use Google's Maven Central mirror
We are using Maven 3.1.1 which by default uses HTTP instead of HTTPS for
resolving artifacts from Maven Central. Maven Central recently
discontinued HTTP support. Therefore the build on Travis started
failing. By using an HTTPS mirror of Maven Central the build on Travis
will work again.

I chose Google's mirror because Travis uses this mirror by default, too.
I did not upgrade to a new version of Maven because there is no newer
version with Java 5 support and it should be possible to build JUnit 4
with Java 5 so that we can easily ensure that it works with Java 5.