Tidy up FailOnTimeoutTest
Improve readability and reusability. When I started to read the test it
took me some time to understand it. The reason was the configurable
statement TestStatement and the evaluateWith... methods. The refactored
test uses fixed Statements when possible and uses smaller methods for
wrapping FailOnTimeout with a ThrowingRunnable and for wrapping an
arbitrary statement with a FailOnTimeout. It also calls
FailOnTimeout#evaluate directly when possible.